Remove pipeline proxies and update imports
