+++ mypy.ini
@@ -3,7 +3,7 @@ python_version = 3.11
 mypy_path = src
 exclude = ^src/pipeline/observability/|^src/cli/templates/|^src/plugins.resources/
 strict = True
-files = src/pipeline/context.py,src/pipeline/manager.py,src/pipeline/tools/execution.py
+files = src/entity/core/context.py,src/pipeline/manager.py,src/pipeline/tools/execution.py
 follow_imports = skip
 
 [mypy-langchain_core.*]
