+++ src/pipeline/__init__.py
@@ -114,12 +114,12 @@ def __getattr__(name: str) -> Any:
         "BaseResource": "pipeline.resources",
         "Resource": "pipeline.resources",
         "PipelineStage": "pipeline.stages",
-        "PluginContext": "pipeline.context",
-        "ConversationEntry": "pipeline.context",
-        "ToolCall": "pipeline.context",
-        "FailureInfo": "pipeline.state",
-        "LLMResponse": "pipeline.state",
-        "PipelineState": "pipeline.state",
+        "PluginContext": "entity.core.context",
+        "ConversationEntry": "entity.core.state",
+        "ToolCall": "entity.core.state",
+        "FailureInfo": "entity.core.state",
+        "LLMResponse": "entity.core.state",
+        "PipelineState": "entity.core.state",
         "BasePlugin": "entity.core.plugins",
         "ResourcePlugin": "entity.core.plugins",
         "ToolPlugin": "entity.core.plugins",
