--- src/pipeline/context.py
@@ -1,4 +1,0 @@
-from entity.core.context import ConversationEntry, PluginContext
-from pipeline.state import ToolCall
-
-__all__ = ["PluginContext", "ConversationEntry", "ToolCall"]
