+++ src/pipeline/initializer.py
@@ -21,7 +21,7 @@ from entity.workflows.discovery import (
     register_module_workflows,
 )
 
-from .base_plugins import BasePlugin, ResourcePlugin, ToolPlugin
+from entity.core.plugins import BasePlugin, ResourcePlugin, ToolPlugin
 from .defaults import DEFAULT_CONFIG
 from .logging import configure_logging, get_logger
 from .stages import PipelineStage
@@ -94,7 +94,7 @@ class ClassRegistry:
     def _type_default_stages(self, cls: type[BasePlugin]) -> List[PipelineStage]:
         """Return default stages for the given plugin class."""
 
-        from .base_plugins import AdapterPlugin, PromptPlugin
+        from entity.core.plugins.base import AdapterPlugin, PromptPlugin
 
         if issubclass(cls, ToolPlugin):
             return [PipelineStage.DO]
@@ -313,7 +313,7 @@ class SystemInitializer:
     def _type_default_stages(self, cls: type[BasePlugin]) -> List[PipelineStage]:
         """Return default stages for the given plugin class."""
 
-        from .base_plugins import AdapterPlugin, PromptPlugin
+        from entity.core.plugins.base import AdapterPlugin, PromptPlugin
 
         if issubclass(cls, ToolPlugin):
             return [PipelineStage.DO]
