+++ src/pipeline/resources/__init__.py
@@ -1,6 +1,6 @@
 """Convenience imports for pipeline resources."""
 
-from .container import ResourceContainer
+from entity.core.resources.container import ResourceContainer
 
 from .memory import Memory  # noqa: F401
 from .llm import UnifiedLLMResource  # noqa: F401
