--- src/pipeline/resources/container.py
@@ -1,3 +1,0 @@
-from entity.core.resources.container import ResourceContainer
-
-__all__ = ["ResourceContainer"]
