+++ src/pipeline/tools/execution.py
@@ -5,8 +5,7 @@ import time
 from typing import Any, Dict
 
 from entity.core.plugins.base import ToolExecutionError
-from entity.core.state import ToolCall
-from pipeline.state import PipelineState, MetricsCollector
+from entity.core.state import ToolCall, PipelineState, MetricsCollector
 from entity.core.registries import SystemRegistries
