+++ src/plugins/builtin/resources/echo_llm.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Any, Dict
 
-from pipeline.state import LLMResponse
+from entity.core.state import LLMResponse
 
 from .llm_resource import LLMResource
