+++ tests/integration/test_full_pipeline.py
@@ -5,8 +5,6 @@ from pipeline import PipelineStage
 from entity.core.plugins import BasePlugin
 
 from entity import Agent
-from pipeline import PipelineStage
-from pipeline.base_plugins import BasePlugin
 
 
 class ParsePlugin(BasePlugin):
