+++ tests/test_cache.py
@@ -1,20 +1,17 @@
 from datetime import datetime
 
-import pipeline.context as context_module
+import entity.core.context as context_module
 from entity.core.resources.container import ResourceContainer
-from pipeline import (
+from entity.core.state import (
     ConversationEntry,
     MetricsCollector,
-    PipelineStage,
     PipelineState,
-    PluginContext,
-    PluginRegistry,
-    SystemRegistries,
-    ToolRegistry,
+    ToolCall,
 )
+from pipeline import PipelineStage, PluginRegistry, SystemRegistries, ToolRegistry
+from entity.core.context import PluginContext
 from entity.core.plugins import PromptPlugin
 from pipeline.cache import InMemoryCache
-from pipeline.state import ToolCall
 from pipeline.tools.execution import execute_pending_tools
 from plugins.builtin.resources.llm_base import LLM
 from user_plugins.resources.cache import CacheResource
