+++ tests/test_execute_pending_tools.py
@@ -2,16 +2,14 @@ import asyncio
 from datetime import datetime
 
 from entity.core.resources.container import ResourceContainer
-from pipeline import (
+from entity.core.state import (
     ConversationEntry,
     MetricsCollector,
     PipelineState,
-    PluginRegistry,
-    SystemRegistries,
     ToolCall,
-    ToolRegistry,
 )
-from pipeline.context import PluginContext
+from pipeline import PluginRegistry, SystemRegistries, ToolRegistry
+from entity.core.context import PluginContext
 from pipeline.tools.execution import execute_pending_tools
