+++ tests/test_state_logging.py
@@ -6,7 +6,7 @@ from datetime import datetime
 from pipeline import PluginRegistry, PromptPlugin, SystemRegistries, ToolRegistry
 from pipeline.pipeline import execute_pipeline
 from pipeline.stages import PipelineStage
-from pipeline.state import ConversationEntry, PipelineState
+from entity.core.state import ConversationEntry, PipelineState
 
 from entity.core.resources.container import ResourceContainer
 from entity.core.state_logger import LogReplayer, StateLogger
