+++ user_plugins/failure/basic_logger.py
@@ -7,7 +7,7 @@ from entity.core.plugins import FailurePlugin
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:  # pragma: no cover - type hints only
-    from pipeline.context import PluginContext
+    from entity.core.context import PluginContext
 from pipeline.stages import PipelineStage
