+++ user_plugins/prompts/chain_of_thought.py
@@ -3,7 +3,8 @@ from __future__ import annotations
 from typing import List
 
 from entity.core.plugins import PromptPlugin
-from pipeline.context import ConversationEntry, PluginContext
+from entity.core.state import ConversationEntry
+from entity.core.context import PluginContext
 from pipeline.stages import PipelineStage
