+++ user_plugins/prompts/complex_prompt.py
@@ -3,7 +3,8 @@ from __future__ import annotations
 from typing import List
 
 from entity.core.plugins import PromptPlugin
-from pipeline.context import ConversationEntry, PluginContext
+from entity.core.state import ConversationEntry
+from entity.core.context import PluginContext
 from entity.resources.memory import Memory
 from pipeline.stages import PipelineStage
